claude/hook_daemon: consolidate kubeconfig module at hook_daemon/ level
The kubeconfig helpers lived at session_start/kubeconfig.py while their
only other caller (write_kubeconfig_cli.py) lived one level up at
hook_daemon/. Move kubeconfig.py and test_kubeconfig.py up to
hook_daemon/ so both halves live together and the import graph is
honest.

Also address Copilot review on #1309:

- wrap the write-kubeconfig subcommand in a try/except that prefixes
  errors with "write-kubeconfig failed:" so the failure mode is
  distinguishable from other hook dispatch errors
- atomic 0o600 file creation in write_kubeconfig_cli: open a dot-tmp
  sibling with O_CREAT|O_TRUNC and mode 0o600 so the file is never
  briefly readable at 0o644, then Path.replace onto the final path
- yaml.safe_dump with sort_keys=False to keep the kubeconfig
  top-level order stable

The other Copilot suggestion (lazy import of write_kubeconfig_cli
inside the subcommand branch) conflicts with ruff PLC0415 and
STYLE.md "Imports at module top", so the import stays at module top.

BAZEL_TEST_INVOCATIONS=buildbuddy:8d3263d6-70a2-4b05-9045-ca09fcf15d40
